Separate BaseVoltage from DefaultDiagramStyleProvider
